Add default profile e2e test case
The e2e tests now verify that the default profile is in place and
contains the right content.

We had to fix an issue where the kubelet path would not be accessible
(`0700`) by the user 2000. This has now being worked around by running
an additional `chmod +x /var/lib/kubelet` on operator boostrap.

Some additional logging has been added to the test framework as well.

Signed-off-by: Sascha Grunert <sgrunert@suse.com>