Add e2e test to ensure operator can be cleaned up and redeployed
A recent bug was indentified that resulted in the operator failing to be
installed again after it was cleaned up because the symlink to the
seccomp profiles directory. It was fixed by checking that the symlink
does not already exist in the directory before creating it in the init
container. This adds an e2e test to make sure that this bug is not
reintroduced.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>