Initialize the operator with a dedicated subcommand
Instead of relying on a bunch of bash code we now write the
initialization phase directly as operator subcommand. This way we can
reuse the already built image and test the code paths.

A new `config` package now contains global consts shared across the
whole application. This way we can decouple it from the `profile`
package.

Signed-off-by: Sascha Grunert <sgrunert@suse.com>