tree 78aba16386ab2a0b5f6d2a5e0033df9b49f45450 parent 9d80d5f22a2ede5aee7d139dc8c82109c4dff781 author Sascha Grunert <sgrunert@suse.com> 1596710731 +0200 committer Sascha Grunert <sgrunert@suse.com> 1596720241 +0200 gpgsig -----BEGIN PGP SIGNATURE-----
 iQFGBAABCAAwFiEEkoNsU4c5ikSa95TPjOAp3RqGblIFAl8sBHESHHNncnVuZXJ0
 QHN1c2UuY29tAAoJEIzgKd0ahm5SJCYH/1c9jWo0QaogolH1Ocp0f2KAGnLRflwp
 eX7cNcnVPyN6Ji+5vDv0iJDwOExDRWfqEp3F/0TGol1lW6zrEfDaMdNa72TlubnE
 l8S6IuttGQn4BFcM5f5TUS+ECESLGBLmOVuvxewhkx+7+IyCeiT+ItBYScr8P+Nb
 qxpVKtOtGaF3usvC7/zDmtHXRMG33Cbu6psD2LCVebRbJGvw5V13y18LwIpbALrc
 d8l6O512CtIhrnMiLH8foMJm074d7bDUSqPSnG2A0MxMXSRAQdgSmCsPMwwlA95E
 r2Hfu0JmMYCio4LiCad3FwRE49MDU/BE+G0XYJIDoDFltNU8apzlSHA=
 =roJW
 -----END PGP SIGNATURE-----

Initialize the operator with a dedicated subcommand

Instead of relying on a bunch of bash code we now write the
initialization phase directly as operator subcommand. This way we can
reuse the already built image and test the code paths.

A new `config` package now contains global consts shared across the
whole application. This way we can decouple it from the `profile`
package.

Signed-off-by: Sascha Grunert <sgrunert@suse.com>