Remove ConfigMap support
With this commit, the controller will no longer watch for seccomp
profiles from ConfigMaps and only watch for SeccompProfile resources.
Remove the old ConfigMaps and related RBAC from the deployment manifests
and remove related e2e tests.

This requires modifications to the default profiles. The
security-profiles-operator profile remains a ConfigMap because the
controller's init container needs to mount and copy it in order to
bootstrap the controller, but the special annotation is no longer
needed. This has the consequence that this particular profile cannot be
managed as a SeccompProfile resource, it can only be managed as a
ConfigMap. The nginx-1.19.1 profile is transformed to a SeccompProfile
resource. Since there is a slight delay between a CRD being registered
with the API server and it being available to POST to, the CRD
definition is split into its own manifest so it can be deployed first
before the default profiles.