tree 727fd778097679bbc85014652e01d763ed3fddda parent b178530ed74aae8619c25a65b2a60169d4b42736 author Colleen Murphy <colleen.murphy@suse.com> 1601571070 -0700 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1603828317 -0700
Remove ConfigMap support

With this commit, the controller will no longer watch for seccomp
profiles from ConfigMaps and only watch for SeccompProfile resources.
Remove the old ConfigMaps and related RBAC from the deployment manifests
and remove related e2e tests.

This requires modifications to the default profiles. The
security-profiles-operator profile remains a ConfigMap because the
controller's init container needs to mount and copy it in order to
bootstrap the controller, but the special annotation is no longer
needed. This has the consequence that this particular profile cannot be
managed as a SeccompProfile resource, it can only be managed as a
ConfigMap. The nginx-1.19.1 profile is transformed to a SeccompProfile
resource. Since there is a slight delay between a CRD being registered
with the API server and it being available to POST to, the CRD
definition is split into its own manifest so it can be deployed first
before the default profiles.