wip: Add new Controller for collecting recorded seccomp profiles
This controller is watching pods carrying the "io.containers.trace-syscall"
annotation which is used by the oci-seccomp-bpf hook to trigger the
tracing of syscalls issued by a pod to generate a seccomp profile.

After pod completion the controller will collect the generated profile
and create a "SeccompProfile" resource in the namespace that the traced
pod was running in.

Signed-off-by: Ralf Haferkamp <rhafer@suse.com>