tree a1d6731398c1dfd1042af870652c92f35ade6f38 parent ad6de5adc308755bb23e2fe95c2c7e235f4a7506 author Colleen Murphy <colleen.murphy@suse.com> 1604367452 -0800 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1604477644 -0800
Add ability to merge seccomp profiles

Add an attribute BaseProfileName to the SeccompProfile resource. When
reconciling a SeccompProfile, the controller will look up a profile with
this name in the same namespace and use it as a base upon which the
syscalls of both resources are unioned. This allows users to be able to
create a base profile for their container runtime and then add only the
syscalls needed for a particular application.

The base profile lookup is not recursive, so only two profiles can be
combined.

The controller only combines syscalls, not flags or architectures or the
default action. It does not validate whether a syscall is specified for
multiple contrary actions.