Fix SeccompProfile status update idempotency
Without this commit, if the operator daemonset is running on more than a
single node and tries to update the status of a SeccompProfile resource,
it will will fail but then incorrectly return success in the reconciler.
This is because the controller is not idempotent and all replicas of the
controller are trying to update the same resource and causing conflicts.
This commit checks the status of the resource before trying to bother
update it so that the error handling is correct and warning events are
emitted when necessary.