tree b67400172fbec12e985eb11572ee158177ce5723 parent 05e7e388b9d76b625355e91590c3a802d8a8b183 author Colleen Murphy <colleen.murphy@suse.com> 1604533317 -0800 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1605685203 -0800
Fix SeccompProfile status update idempotency

Without this commit, if the operator daemonset is running on more than a
single node and tries to update the status of a SeccompProfile resource,
it will will fail but then incorrectly return success in the reconciler.
This is because the controller is not idempotent and all replicas of the
controller are trying to update the same resource and causing conflicts.
This commit checks the status of the resource before trying to bother
update it so that the error handling is correct and warning events are
emitted when necessary.