Add finalizers to delete SeccompProfiles
This patch adds the ability to delete a seccomp profile from disk when
its corresponding SeccompProfile resource is deleted. It does this by
adding finalizers corresponding to the node the controller is running on
to each SeccompProfile as it is reconciled, and deleting the profile
file and the finalizer upon resource deletion.

Allowing the operator to delete a file under the the
'/var/lib/security-profiles-operator' symlink required adding the
'unlinkat' syscall to its own seccomp profile. The operator also now
requires RBAC permissions to update a SeccompProfile.

In order to ensure uninstallation with `kubectl delete -f` works
properly, this changes the kustomization configuration and installation
instructions so that default profiles are created separately from the
rest of the operator. This allows them to be cleaned up deliberately by
the user instead of having the controller try to delete them before
being terminated and before losing its RBAC permissions. Without this,
if the operator is deleted with `kubectl delete -f operator.yaml` the
namespace and profiles will be stuck in a "terminating" state because
the pods running the controller are unable to finish processing their
deletion.

To prevent profiles in use from being deleted and breaking the pods
using them, a new field ActiveWorkloads is added to SeccompProfileStatus
to link profiles to the pods using them, and ensure a finalizer is added
to prevent an active profile from being deleted. Also add a Status field
within the Status object to flag the profile as "Active" or
"Terminating" if there are still finalizers on the profile preventing it
from being removed.