Integrate selinux-operator into the SPO
This integrates the selinux-operator [1] into the SPO by copying in the
reconcilers/controllers and setting them up in the code-base.

Note that since the selinux-operator was meant to run as part of a
deployment, I had to create two managers in the main function. One for
the daemonset that handles the seccomp profiles, and another one for the
Selinux controllers which will do leader election, and thus effectively
only run one.

[1] https://github.com/JAORMX/selinux-operator