tree d9b86d963df0f03cbc41ffbdef8fb0b88be7e8c1 parent 3d60c08eb189aa907903cebccf3d70812531b0da author Juan Antonio Osorio Robles <jaosorior@redhat.com> 1607432905 +0200 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1607604294 -0800
Integrate selinux-operator into the SPO

This integrates the selinux-operator [1] into the SPO by copying in the
reconcilers/controllers and setting them up in the code-base.

Note that since the selinux-operator was meant to run as part of a
deployment, I had to create two managers in the main function. One for
the daemonset that handles the seccomp profiles, and another one for the
Selinux controllers which will do leader election, and thus effectively
only run one.

[1] https://github.com/JAORMX/selinux-operator