tree ee5e3ec442fff4b4b797c8594c37c16175a09128 parent e4c133f24f8bb03fb784bbc818791443f035b45e author Colleen Murphy <colleen.murphy@suse.com> 1607467116 -0800 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1607585813 -0800
Add localhostProfile attribute and column

Add a new attribute, `localhostProfile`, to the SeccompProfile CRD and
ensure it is exposed via kubectl's `-o wide` formatting option. This
attribute represents the string that a user would have to supply to
their container, pod, or application (Deployment, ReplicaSet, DaemonSet,
etc) manifest in order for their application to use the profile. This
makes it clearer than just checking the Path attribute, since it's not
obvious that the `/var/lib/kubelet/seccomp/` part of the path is not
supposed to be included in the pod spec.