Add localhostProfile attribute and column
Add a new attribute, `localhostProfile`, to the SeccompProfile CRD and
ensure it is exposed via kubectl's `-o wide` formatting option. This
attribute represents the string that a user would have to supply to
their container, pod, or application (Deployment, ReplicaSet, DaemonSet,
etc) manifest in order for their application to use the profile. This
makes it clearer than just checking the Path attribute, since it's not
obvious that the `/var/lib/kubelet/seccomp/` part of the path is not
supposed to be included in the pod spec.