tree ba9e0124d1f74df2f07da4f5b11f76d6107831de parent 5cfb796d4781449bf8bf3ae7af76f4ef6171cce0 author Juan Antonio Osorio Robles <jaosorior@redhat.com> 1607956946 +0200 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1608033229 -0800
Enable e2e tests to run on OpenShift

This changes the e2e tests to not only run on a kind cluster, but to
also be able to run on a running OpenShift cluster.

To enable this, there is now the `E2E_CLUTER_TYPE` environment variable
which the Makefile can take into use and that will be taken into account
by the test itself.

Note that in order for the security-profiles-operator to work on
OpenShift, I temporarily added spc_t as the SELinux type. This ensures
that the operator is allowed to place files on the folder it's trying to
access. This is a placeholder while we work on a better mechanism to
consume both workload and bootstrap SELinux policies [1]. This mechanism
will eventually replace the tiny workloads that we currently spawn to
install policies and will be included as part of the main daemonset.

[1] https://github.com/JAORMX/selinuxd