tree 51c56c25a3a3a4239bbf9b8914ee49c52278a12e parent 6842f932dbc4be37b857773f328230ae3c31eeab author Colleen Murphy <colleen.murphy@suse.com> 1607973621 -0800 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1611304407 -0800
Add ProfileBinding CRD and webhook

Add a new CRD called ProfileBinding which defines a relationship between
a container image and a profile resource (currently only
SeccompProfile). When a pod that has a matching container image is
created, a mutating admission webhook injects a SecurityContext into it
to refer to the SeccompProfile. This change adds a new controller and
webhook as a separate Deployment. The webhook adds a new dependency on
cert-manager to manage the certificates it uses to register with the API
server.