Add retries for setting SeccompProfile status
Currently, the seccomp controller retries its calls to update the
finalizer list for SeccompProfiles. This is because in a multi-node
deployment, the operator DaemonSet has no elected leader and so each
node attempts to perform an update via the API and all except the first
one fail, so retrying ensures each node can get a refreshed copy of the
resource and perform the update safely. Without this patch, other types
of resource updates, such as to add an attribute to the resource status,
are not retried. In the case of adding a path attribute to the resource
status, there is no consequence to failing the update since it is the
same data for all nodes. However, the failed attempt leaves a misleading
error in the logs. In the case of adding a link to an active workload,
there is a mild consequence of not having a full accounting of all
active linked pods in the SeccompProfile resource, which would have been
rectified on the next reconciler loop. This patch adds retries to the
rest of the resource status updates. This will only cause additional API
calls in the event the resource does need to be updated, otherwise it
will noop.