tree 07695144f1ff8ee04636fdbc77e77d4d615c42cb parent 5e06b7842ed289f0793985c7fc6c7649a8cfff3f author Juan Antonio Osorio Robles <jaosorior@redhat.com> 1608121250 +0200 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1608196390 -0800
Don't remove SelinuxPolicy finalizer until the ConfigMap is deleted

This effectively makes the SelinuxPolicy wait until the CM has been
deleted... We were having the case that if the pods took too long to
delete, the selinux policy object would be deleted but the policy would
still be there as the pods were still waiting for deletion.

This fixes that and now the state of the object accurately matches the
state of the cluster

In order to make this effective, the "foregroundDeletion" finalizer was
added to the policy ConfigMap. This ensures that it won't be deleted
until the dependents are done.