Don't remove SelinuxPolicy finalizer until the ConfigMap is deleted
This effectively makes the SelinuxPolicy wait until the CM has been
deleted... We were having the case that if the pods took too long to
delete, the selinux policy object would be deleted but the policy would
still be there as the pods were still waiting for deletion.

This fixes that and now the state of the object accurately matches the
state of the cluster

In order to make this effective, the "foregroundDeletion" finalizer was
added to the policy ConfigMap. This ensures that it won't be deleted
until the dependents are done.