Reorganize CRD type definitions
controller-gen is nondeterministic about the order in which it generates
YAML manifests from Go types, so `make manifests` may put them in a
different order and cause unnecessary changes to the generated YAML
files. The code is growing to a point that a little reorganization is
nice anyway. This change splits the types into their own packages so
that the Makefile command can have more fine-grained control over which
types to generate in which order.