tree f25988a6f4559cb2c7577daddf54bce2ef102fce parent 83ea1bf4b55a9eea3bb0484a324edd01c59d4da1 author Colleen Murphy <colleen.murphy@suse.com> 1608237124 -0800 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1608276145 -0800
Reorganize CRD type definitions

controller-gen is nondeterministic about the order in which it generates
YAML manifests from Go types, so `make manifests` may put them in a
different order and cause unnecessary changes to the generated YAML
files. The code is growing to a point that a little reorganization is
nice anyway. This change splits the types into their own packages so
that the Makefile command can have more fine-grained control over which
types to generate in which order.