Change controller manager to be a Deployment
Why?
====

While the current approach of deploying a DaemonSet directly works, it
has some limitations:

* It's not trivial to customize what to deploy (e.g. if we want to
  disable AppArmor)

* It's not directly supported by some operator catalogs (e.g.
  via the Operator Lifecycle Manager)

This instead proposes the usage of a Deployment.

This Deployment listens for a Configuration called "config", which
will allow us to customize the DaemonSet (e.g. enable/disable
features).

On the other hand, with we'll be able to use this Deployment to serve
webhooks (like the one that's currently missing to validate the SELinux
policy namespaces).

What's missing?
===============

Actually add configuration options
----------------------------------

Currently the controller only deploys the DaemonSet... no configuration
is taken into account.

Handle configuration updates
----------------------------

Once we have configurations, we need to handle updates.

Better bindata management
-------------------------

Currently the DaemonSet is a hardcoded string, it would be better to
have an appropriate bindata solution which would read a yaml file and
update go code to handle it automatically.

Merge SELinux to the DaemonSet
------------------------------

Currently SELinux works on a cluster level, and so it's part of the main
Deployment. This needs to be changed so SELinux module installation
should be done as part of the DaemonSet instead. This will require us to
change the inner workings of the SELinux bits (probably using selinuxd
[1])

[1] https://github.com/JAORMX/selinuxd