Take selinuxd into use
Changes SELinux integration by ditching the intermediate configMaps
previously used by the CM reconcile loop of the selinux controller.
Instead, the daemon running in the DS now takes an boolean argument that
signifies if the SELinux feature should be enabled. This is false by
default.

If true, the DS deploys an additional container running selinuxd which
listens to any files under /etc/selinux.d and can be queried for status
over a UNIX socket mounted under /var/run/selinuxd. Both the socket
directory and the drop-in directory are EmptyDirs shared between
selinuxd and the controller.

In addition, the operator receives a --with-selinux=true flag, which
makes the operator create an additional manager that listens for the
selinuxpolicy API object. Contents of those objects are reconciled to
the /etc/selinux.d drop-in directory and then the controller queries
the state of selinuxd to make sure the policy was indeed added or
removed.