tree 5accdbab8f04d9cd46f1d8a2926aba0852840290 parent c7f337312e70b657af8e8c51b9310c47ea74f2a7 author Juan Antonio Osorio Robles <jaosorior@redhat.com> 1610538259 +0200 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1610542235 -0800
Add Makefile targets for easy development in OpenShift

This adds the following targets:

`push-openshift-dev`
--------------------

Builds the operator image based on UBI, and pushes it to the OpenShift
cluster that your current kubectl context is pointing at.

`do-deploy-openshift-dev`
-------------------------

Builds `deploy/operator.yaml` based on an `openshift-dev` overlay which
contains overrides that point to the local OpenShift container image
registry. This is then deployed and a trigger is set so any changes to
the container image get automatically taken into use by the Deployment.

`deploy-openshift-dev`
----------------------

Triggers both of the aforementioned targets.

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>