Fix UBI Dockerfile
A recent commit added the seccomp tag to the build [1], however, it
missed to add it to the UBI build. This fixes that. Note that since the
seccomp-static package is not available in UBI, we switched the
build-base to be fedora, and it'll eventually copy the binary to a UBI
based container.

[1] f504483024c68b621f6e43170de59e1cd2f47216

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>