tree f21c3a5079e7d37b11652fb8bf2ebdd4061fee47 parent f504483024c68b621f6e43170de59e1cd2f47216 author Juan Antonio Osorio Robles <jaosorior@redhat.com> 1610966518 +0200 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1610969864 -0800
Fix UBI Dockerfile

A recent commit added the seccomp tag to the build [1], however, it
missed to add it to the UBI build. This fixes that. Note that since the
seccomp-static package is not available in UBI, we switched the
build-base to be fedora, and it'll eventually copy the binary to a UBI
based container.

[1] f504483024c68b621f6e43170de59e1cd2f47216

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>