Separate RBAC permissions for spo-manager and spod
This separates the RBAC permissions for the SPO manager and the SPO
daemon. They now have the following service accounts with matching roles
and cluster roles:

* spo-manager
* spod

They are managed by the `make generate` target which, in turn, calls
`controller-gen` to generate the necessary RBAC roles. These are set as
comments that kubebuilder reads.

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>