tree 1def80638eacd769b1e7fd6c5e248dafdefa12ea parent 862fe1a3f014b162d890963f42947ec2975cdfbc author Jakub Hrozek <jhrozek@redhat.com> 1611149646 +0100 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1611323369 -0800
tests: Work around race condition in selinuxpolicy status by retrying the is-installed/is-removed check

The selinuxpolicy API object is a singleton and has a single finalizer
and a single status. This means that the first instance of selinuxd that
removes the finalizer or sets the status 'wins' as far as anyone waiting
for status is concerned, but there's no guarantee that the other pods in
the DS running on the other nodes have finished their work as well.

To work around that until we address the issue systematically, let's
retry the check on the nodes up to three times with a sleep of 5 seconds
between the tries.