tree 67d96c8607b3d97131c979f8b27ad366c6e72993 parent ea7c801e603db356e7713040b7547f3720769e7b author Jakub Hrozek <jhrozek@redhat.com> 1611149954 +0100 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1611323369 -0800
tests: Skip building images when testing on OpenShift if E2E_SKIP_BUILD_IMAGES is set

When iterating on the tests, the test driver would still be building the
image which is irritating. This patch amends the tests so that if the
E2E_SKIP_BUILD_IMAGES is set to a value that evaluates to true, the
images are not rebuilt again.