selinux: Wait for the /ready endpoint before writing the policies, write policies only if they differ
It was pointed out during the initial review of the selinuxd integration
that the selinuxpolicy controller performs too many needless writes on
each reconcile while it's waiting for selinuxd to do its job and install
a policy.

Just naively perform a single write uncovered a race in selinuxd where
the initial inotify messages might have gotten lost before selinuxd was
done processing the initial shipped policies. To mitigate that, selinuxd
has now a /ready handler that exposes whether it is ready to process
policies.

This patch avoids reconciling any policies until the /ready handler
replies True. Making use of this, a policy is only written if it doesn't
exist yet or if the contents differ.