tree 1a6774f456b40ad3f783ace8c96e05b0d039af0f parent b432ee4fa583acd71c7dd89531513889442ffa80 author Jakub Hrozek <jhrozek@redhat.com> 1611335993 +0100 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1611746020 -0800
selinux: Wait for the /ready endpoint before writing the policies, write policies only if they differ

It was pointed out during the initial review of the selinuxd integration
that the selinuxpolicy controller performs too many needless writes on
each reconcile while it's waiting for selinuxd to do its job and install
a policy.

Just naively perform a single write uncovered a race in selinuxd where
the initial inotify messages might have gotten lost before selinuxd was
done processing the initial shipped policies. To mitigate that, selinuxd
has now a /ready handler that exposes whether it is ready to process
policies.

This patch avoids reconciling any policies until the /ready handler
replies True. Making use of this, a policy is only written if it doesn't
exist yet or if the contents differ.