tree 3c18a9105866c10a6e319feba4410913f70b330e parent 996e613786f54988b6e50f23c3c9eb68b1443a28 author Jakub Hrozek <jhrozek@redhat.com> 1611604532 +0100 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1611648874 -0800
make: Set EnableSelinux=true through kustomize for OpenShift

When deploying on OpenShift, we are typically interested in deploying
the SELinux controller as well. Since we already have a
OpenShift-specific Makefile target that uses a kustomize overlay,
let's just extend that overlay to change the default values of the
'config' ConfigMap.