tree e7ed7b151561faaff09fc88a2926d350c617b7ed parent cc4e2e8ab5c073e9b0bd661ceef3a72b265663e6 author Jakub Hrozek <jhrozek@redhat.com> 1611606191 +0100 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1611648154 -0800
make: Deploy cert-manager before deploying the operator in OpenShift

There exists the 'deploy-openshift-dev' Makefile command that deploys the
operator in an OpenShift cluster. But since the cert-manager dependency
was added, the target didn't work as OpenShift doesn't come with
cert-manager by default. This patch adds the dependency and waits for
the cert-manager pods to come up.