tree 58f98ae994df849afb5a13a81c57d4808a7dd842 parent b47a76284092f68f72df6c0482e19c80e060d8b0 author Colleen Murphy <colleen.murphy@suse.com> 1611689289 -0800 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1611736180 -0800
Split out ProfileBinding webhook

The ProfileBinding CRD isn't required for the minimum functionality of
the operator, and it requires installing cert-manager additional
dependency which may not be desireable to all users. Furthermore,
cert-manager is causing intermittent problems in the e2e tests that need
to be worked out. This change splits the webhook into its own command
and its own deployment, and skips deploying and testing it in the e2e
tests. The webhook e2e test can still be run manually by setting the
environment variable E2E_RUN_EXPERIMENTAL=Y.