tree 44c2e6683432713fe550a99ab56fbb700ae8a72e parent c665391a5e470210a062d1f92faece944e5a343d author Colleen Murphy <colleen.murphy@suse.com> 1611776964 -0800 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1611907541 -0800
Add wait condition for webhook cert

The cert-manager Deployment seems not to do leader election and so might
need a few tries to create a Secret for a Certificate, which can be seen
in the pod logs:

  E0127 19:41:53.078168       1 controller.go:158] cert-manager/controller/CertificateReadiness
  "msg"="re-queuing item due to error processing" "error"="Operation cannot be fulfilled on
  certificates.cert-manager.io \"webhook-cert\": the object has been modified; please apply your
  changes to the latest version and try again" "key"="security-profiles-operator/webhook-cert"

This affects the security-profiles-operator webhook Deployment's ability
to mount and use the cert. This change ensures the secret is ready to
use before trying to use the ProfileBinding webhook.