tree 14c81c12c4967c54cf461ece5b1a91f42b278f72 parent a5393f869690ed25b31003b4da7675608fbacf7d author Juan Antonio Osorio Robles <jaosorior@redhat.com> 1611819189 +0200 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1612356628 -0800
e2e: Run tests as sub-tests

Currently, the tests e2e are being from a single `TestSuite` which
iterates through the test definitions calling their respective
functions.

However, if one of these definitions were to fail, the error is reported
towards the upper `TestSuite` test.

This uses testify's `Suite` object ability to run sub-tests (similar to
what `testing.T` provides) and instead uses that.

This way, any errors coming from a sub-test will be reported for that
sub-test and it'll be easier to debug.

One thing to note is that given our usage of testify's `Suite`, we
aren't able to do parallel testing in our e2e tests. In order to enable
this, we'd need to move away from the `Suite` object or use a different
framework.

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>