tree a7d38333764581b3b542097cf5bfb812061b309b parent 27277d14de19cbcef0368d004a27c6978d21bd22 author Sascha Grunert <mail@saschagrunert.de> 1612171519 +0100 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1612269028 -0800
Add seccomp profile recording feature

This is the initial implementation of the seccomp profile recording
feature. Right now it only works with CRI-O newer than v1.20.0 as well
as the deployed webhook. Another requirement is a working [OCI seccomp
BPF hook][0], whereas CRI-O has to be configured to watch the hooks'
directory. For example with a drop-in configuration like this:

```
> cat /etc/crio/crio.conf.d/03-hooks.conf
[crio.runtime]
hooks_dir = [
    "/path/to/oci/hook/dir",
]
```

The hook itself consists of a JSON which contains the execution
instructions as well as the path to the hook binary.

To be able to let the webhook do anything useful, we now deploy an
additional CRD called `ProfileRecording`. This CRD only contains a
selector, which will match any label for the application to be recorded.
If we create the workload, the webhook will take care of adding the
correct annotation to the pod. The hook will then record the seccomp
profile into a temporary JSON file. If the workload has been finished,
the new recorder controller will take the JSON file and reconcile it
into a `SeccompProfile` CRD object, which then gets synchronized on
disk.

The actual implementation is pretty rough and not well tested.
Documentation, test enhancements and stability improvements will follow
in later patches.

[0]: https://github.com/containers/oci-seccomp-bpf-hook

Signed-off-by: Sascha Grunert <mail@saschagrunert.de>