Automatically detect if 'podman' should be used.
If no `CONTAINER_RUNTIME` variable is given in the Makefile, this will
try to verify if the podman binary is available; if it is, I'll try to
use it.

The podman binary is explicitly checked as the `podman` package
contains already a `docker` shim, so in both cases (docker or podman)
there will be a `docker` executable available. If the `podman` binary
isn't available, it falls back to docker as a default.

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>