tree d3fa84874524e233708d81cfcefeff149815b5c4 parent 5ec7b22f5a094fb0803740ec529e05448caf0868 author Juan Antonio Osorio Robles <jaosorior@redhat.com> 1612266110 +0200 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1612345828 -0800
Automatically detect if 'podman' should be used.

If no `CONTAINER_RUNTIME` variable is given in the Makefile, this will
try to verify if the podman binary is available; if it is, I'll try to
use it.

The podman binary is explicitly checked as the `podman` package
contains already a `docker` shim, so in both cases (docker or podman)
there will be a `docker` executable available. If the `podman` binary
isn't available, it falls back to docker as a default.

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>