Remove nodeSelector from deployment and webhook
The `nodeSelector` value `node-role.kubernetes.io/master` seems to be
only set by nodes which are created via kubeadm (and others). If the
node does not have the label at all, then the operator will not be
scheduled there. This can be reproduced by the usage of the
`hack/local-up-cluster.sh` script within k/k.

I think we should be safe removing the selector since we already have
the tolerations in place.

Signed-off-by: Sascha Grunert <mail@saschagrunert.de>