tree 390d5ed90275eb1251807f0ee04a8f6e70518142 parent 053ea5805896a189dab8dcb2a4b961a8926469cb author Sascha Grunert <mail@saschagrunert.de> 1612347480 +0100 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1612348348 -0800
Remove nodeSelector from deployment and webhook

The `nodeSelector` value `node-role.kubernetes.io/master` seems to be
only set by nodes which are created via kubeadm (and others). If the
node does not have the label at all, then the operator will not be
scheduled there. This can be reproduced by the usage of the
`hack/local-up-cluster.sh` script within k/k.

I think we should be safe removing the selector since we already have
the tolerations in place.

Signed-off-by: Sascha Grunert <mail@saschagrunert.de>