Add personas and user story documents
These came from our original discussion in a document [1], and are meant
to represent the actual people that this operator would help (the
personas) and the ways that the operator would help these people (as
user stories).

Thus, for future features added to this operator, they should be aimed
at helping one of the personas listed, or at improving or addressing one
of the user stories listed. New personas and user stories can be added,
and doing so will help clarify new functionality proposed to the
operator.

A section to the GitHub templates has been added to include references
to the user stories and personas.

Co-Authored-By: Paulo Gomes <pjbgf@linux.com>
Co-Authored-By: Sascha Grunert <mail@saschagrunert.de>
Co-Authored-By: Colleen Murphy <colleen.murphy@suse.com>
Co-Authored-By: Jakub Hrozek <jhrozek@redhat.com>
Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>

[1] https://docs.google.com/document/d/1x0kS1Mzxcw0k3aq_27QK0OBHSea6apD-WMVwURYZsU4/edit?usp=sharing

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>