tree cdfaf9183d9ba5ffdbd9f935d6aa2136d27a423f parent a54f35bac03bc01be7391ff3f93e9245099fce2b author Juan Antonio Osorio Robles <jaosorior@redhat.com> 1612512189 +0200 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1612856005 -0800
Add personas and user story documents

These came from our original discussion in a document [1], and are meant
to represent the actual people that this operator would help (the
personas) and the ways that the operator would help these people (as
user stories).

Thus, for future features added to this operator, they should be aimed
at helping one of the personas listed, or at improving or addressing one
of the user stories listed. New personas and user stories can be added,
and doing so will help clarify new functionality proposed to the
operator.

A section to the GitHub templates has been added to include references
to the user stories and personas.

Co-Authored-By: Paulo Gomes <pjbgf@linux.com>
Co-Authored-By: Sascha Grunert <mail@saschagrunert.de>
Co-Authored-By: Colleen Murphy <colleen.murphy@suse.com>
Co-Authored-By: Jakub Hrozek <jhrozek@redhat.com>
Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>

[1] https://docs.google.com/document/d/1x0kS1Mzxcw0k3aq_27QK0OBHSea6apD-WMVwURYZsU4/edit?usp=sharing

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>