tree 9bf92bd6fa1868978ae4e1b9bfbd060b16bebdc4 parent 21fc8be846d3ab534569645947619efec312e69c author Juan Antonio Osorio Robles <jaosorior@redhat.com> 1613030281 +0200 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1613033776 -0800
Add conditional flags to run Seccomp and ProfileBinding test

This introduces the following environment variables:

* `E2E_TEST_SECCOMP`
* `E2E_TEST_PROFILE_BINDING`

As the name suggests, these enable/disable the tests pertaining to those
features.

These are turned on by default, but are turned off for the fedora-based
e2e test in order to not repeat tests.

The SELinux flag was renamed to match that convention.

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>