Add conditional flags to run Seccomp and ProfileBinding test
This introduces the following environment variables:

* `E2E_TEST_SECCOMP`
* `E2E_TEST_PROFILE_BINDING`

As the name suggests, these enable/disable the tests pertaining to those
features.

These are turned on by default, but are turned off for the fedora-based
e2e test in order to not repeat tests.

The SELinux flag was renamed to match that convention.

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>