tree 388b1798364d174f594e30534fd61374608b20fe parent 4fb514709f381a38d1db2bd1988d89e68ea6d32d author Juan Antonio Osorio Robles <jaosorior@redhat.com> 1613034658 +0200 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1613047695 -0800
e2e: Use fedora-based selinuxd image for e2e tests

The default image is CentOS-based, which works in CentOS, RHEL and
RHCOS. But our e2e tests use Fedora, which doesn't recognize the
libsemanage configuration format (and most likely the policy format
too). So, let's switch to a fedora-based image for this test.

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>