tree 8af474c342699ae4ebf52d8b6a62b9ad60dc3a83 parent 5d9de0fb65e22beba11fe16b1755d18df3f4d11e author Juan Antonio Osorio Robles <jaosorior@redhat.com> 1613047858 +0200 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1613052495 -0800
Remove clusterRole and clusterRoleBinding overrides from namespaced-operator

The main target of this type of deployment is to have the SPO to only
liten for policies in specific namespaces. Having these overrides only
adds extra complexity, and the operator itself is able to function
without these. Let's remove this extra complexity!

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>