Stick to x86_64 nix binaries for image builds
We cross-compile via nix, so there is no need to use the `$(uname -m)`,
which will cause the `docker buildx` builds to download the arm binaries
instead of the x86_64 ones. This will result in cache poisoning on
cachix and will make the CI fail due to a timeout on the master branch.

Signed-off-by: Sascha Grunert <mail@saschagrunert.de>