tree 568b2821d35f4e52d74209d61743b69b2ae7f0ce parent f0f26afa1b90a99b26cb78ca03c9cd89aee2b825 author Sascha Grunert <mail@saschagrunert.de> 1613554444 +0100 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1613556426 -0800
Stick to x86_64 nix binaries for image builds

We cross-compile via nix, so there is no need to use the `$(uname -m)`,
which will cause the `docker buildx` builds to download the arm binaries
instead of the x86_64 ones. This will result in cache poisoning on
cachix and will make the CI fail due to a timeout on the master branch.

Signed-off-by: Sascha Grunert <mail@saschagrunert.de>