tree 0dfbd523a678ac01ec1c60980bb82579fbc3bce4 parent 7d238624ebdbcbdb33fdb7f0f55145b1a1c53057 author Juan Antonio Osorio Robles <jaosorior@redhat.com> 1613741036 +0200 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1614686599 -0800
Separate RBAC for daemon and deployment

this uses different roles and service accounts for the manager
deployment and the spod. These roles are generated through kubebuilder
annotations, and thus the mechanism and controller-gen calls to do so
were added to the Makefile.

The code was divided in such a way that it's easy to do these
controller-gen invocations and generate the needed rbac rules for the
two separate components. And so, everything that runs on the daemon is
now under the `daemon` directory. The same was done for the manager.

Note that the webhook is still not hooked to this mechanism and that'll
be handled separately.

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>