tree 7debfecf530955bdcbded9ee9bdab401b0cbc9b6 parent 8deda397d8a4f980698e6d3aa89dc454c6ff7da5 author Jakub Hrozek <jhrozek@redhat.com> 1614262570 +0100 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1614681199 -0800
spod: ReadyOnlyRootFs for the selinux container

Recent releases of selinuxd allow to specify a custom DB dir path. We
can take this into use by putting the DB into an EmptyDir where we
alrady have the selinuxd socket. This allows us to run the selinuxd
container with a read-only FS.