tree d163457cbdaa3e32b127c2516b561cf236b32526 parent 637d00de79fcb9936994e6aedcbcc0628fda5322 author Sascha Grunert <sgrunert@redhat.com> 1614697825 +0100 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1614861368 -0800
Switch to latest CRI-O master for e2e tests

To be able to write a test for multi-container recording, we need the
latest CRI-O running within the e2e test suite. This also enables the
usage of the latest crun release, which is now per default available
within the bundle.

Signed-off-by: Sascha Grunert <sgrunert@redhat.com>