tree 0d0ce726a5d1d4602fd832d5e1ac50bec73a6294 parent 0507c4e6f7646ec163f36abbfcc4d2c9de381c0e author Sascha Grunert <sgrunert@redhat.com> 1614872080 +0100 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1614877569 -0800
Add multi-container recording support

This adds support for multi container recording via the OCI seccomp BPF
hook. To not confuse users, we now require the (unreleased) version of
CRI-O v1.21.0. We always suffix the recording by the container name,
which allows the user to distinguish the profiles in a sane way.

Signed-off-by: Sascha Grunert <sgrunert@redhat.com>