tree 4ebf175f15d6f2566daa03cf6c87d0a05b980fe9 parent 232d14b77146d58b8b3bb3d221168ec3b7dd1b58 author Jakub Hrozek <jhrozek@redhat.com> 1614873886 +0100 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1614877575 -0800
Re-enable SELinux by default on OCP

Prior to using CRDs, we were using kustomize to change the enableSelinux
key of the config CM to true. This no longer works since we switched to
CRD as it appears that kustomize doesn't handle CRD:
    https://github.com/kubernetes-sigs/kustomize/blob/1d4b3fa36c8918ad884c7aaa6a055fa2d0fa9bf9/examples/transformerconfigs/crd/README.md

Let's re-enable SELinux by default on OpenShift by just using sed.
Long-term we want to switch to auto-detect the capabilities anyway.