spod: Lower the selinuxd CPU requests and limits
The CPU requests and limits were way too high for the selinuxd container,
which was causing the DS pods to not be scheduled on clusters with low
CPU core nodes.

This time I did better experimenting with the different limits --
setting anything more than 1000m of CPU doesn't really make sense as a
limit, because rarely there is more than one process executing (always
selinuxdctl, sometimes there's an added setfilecontext helper).

With different CPU limits the selinuxd initialization took, until the
READY signal approximately:
    - with 300m CPU: 9 minutes
    - with 500m CPU: 6 minutes
    - with 1000m CPU or more: about 3 minutes
So the times scale linearly, but there's really no point in setting a
full core request (the process runs fine with much less) and no point in
setting a double-core limit (the process never uses them).

With memory, we unfortunately need to stick to the high values, I tested
setting request to 256Mi but got OOMKilled. So maybe 384 might work as
the lowest request, but I'm afraid it might not with complex policies.