tree 8313e9b4c101ec3dd18e8cff0c00217cce6df504 parent 8317d76ed92d74249aa22b1eab06859546218e1e author Jakub Hrozek <jhrozek@redhat.com> 1615211766 +0100 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1616057359 -0700
Per-node statuses

This PR introduced a new API secpolnodestatus that aims to address the
following problems:
 - it was impossible to reflect per-node status such as a profile failed
   to install on a single node or a node does not support the selected security
   profile
 - simply adding per-node attributes like a map might not scale, given
   very high number of nodes, the object might become too big for etcd's 1MB
   limit (see also kubernetes/kubernetes#73324)
 - because the per-profile status is written to by several sources (each
   pod in a DaemonSet), the status might appear as "flapping" as different pods
   reach different states at their own pace.

The secpolnodestatus is created, managed and deleted together with finalizers
through an API in a new module. Instead of updating the global state directly,
the DS pods now call the API update method that updates the node status and if
needed, also the global status.

When a policy is deleted, the object is marked as terminating, when the policy
payload is removed, the node status object along with its finalizer is deleted.
Finally, when all finalizers are gone, so is the global policy object.