tree 607a7895bad671f65a09e85f412f28ccc02f76df parent f72243f01f6f792b54febfb5d6c67c6deaeed21a author Jakub Hrozek <jhrozek@redhat.com> 1615557269 +0100 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1615817106 -0700
tests: Enable SELinux in tests

Since the recent change to create spod in the operator itself, the spod
resource has been created with selinux disabled, thus failing all the
tests.

Long-term we should probably auto-detect the platform capabilities, but
for now, let's just enable selinux if selinux-related tests are selected
and wait for the DS to roll out.