tests: Enable SELinux in tests
Since the recent change to create spod in the operator itself, the spod
resource has been created with selinux disabled, thus failing all the
tests.

Long-term we should probably auto-detect the platform capabilities, but
for now, let's just enable selinux if selinux-related tests are selected
and wait for the DS to roll out.