Retry with backoff in the controllers
The retry logic in the controllers was often suboptimal in the sense
that a) retries followed immediatelly after the first failure and b)
sometimes the retry didn't re-fetch the original resource.

This patch aims to make the retry logic more robust. Longer-term, we
should just get rid of retries and gracefully reconcile again instead.

Because the new Retry handler typically also Gets the resource to
refresh it, both IsConflict and IsNotFound errors are ignored.