tree d22fa26e8e60546c428aa5d98f687b7529b7ae03 parent 809cae9c336685ecdca636c17c9df296f1ec648b author Jakub Hrozek <jhrozek@redhat.com> 1615818006 +0100 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1616057359 -0700
Retry with backoff in the controllers

The retry logic in the controllers was often suboptimal in the sense
that a) retries followed immediatelly after the first failure and b)
sometimes the retry didn't re-fetch the original resource.

This patch aims to make the retry logic more robust. Longer-term, we
should just get rid of retries and gracefully reconcile again instead.

Because the new Retry handler typically also Gets the resource to
refresh it, both IsConflict and IsNotFound errors are ignored.