tree e7e3647db48b04cb43c3b27edbe8db277a389c08 parent c4b28c17a5c29492bcf64595f77741beda98e003 author Jakub Hrozek <jhrozek@redhat.com> 1615975828 +0100 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1616057359 -0700
seccomp: Use the Pending state for seccomp, too

The seccomp controller is very fast which is not bad in itself, but it
presents a challenge in figuring out the state of the seccomp profile.
In some cases, on pod in a DS would be very fast and marked as Installed
while another pod would not even be reconciled yet. This might trigger
the transition of the seccomp profile to Installed as the seccomp
controller would think that there is only one pod in the DS and it's
installed.

This patch addresses the issue somewhat by adding a Pending state with a
short reconcile delay that would hopefully allow all the pods in a DS to
at least pick up the policy which would in turn allow better
synchronization of the global status.

In the meantime, a more systematic solution has been proposed in a
separate issue.