tree 65d9a84b0333334583c6a799d0b36138e7a09463 parent 457c85fdf2c5f9546066115bdb4bfcffd0fb816b author Jakub Hrozek <jhrozek@redhat.com> 1616415604 +0100 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1617817023 -0700
Introduce nodestatus controller

This introduces a new controller called nodestatus controller.

This listens for security profile node status objects and uses those
to decipher what the state of the actual policy is.

As a side note, this also fixes the fact that we were unable to delete
policies if the base one was missing.

We don't really care about the profile being valid or not if the profie
is being deleted, so this moves the logic of deletion above our
validations.

This also separates the PodReconciler to its own controller in the manager

We don't need to track the pods that use the profile in the daemonset,
so this moves it to the controller manager as a workload-annotator
controller.

This ensures that only the Deployment instance updates the workload
itself.

On the other hand, this decreases the reconcile wait time for the
seccomp profile daemon.

Finally, this adds support for multi-namespace listening

The manager was hardcoded to only listen to the operator's namespace;
This is no longer applicable as we do profile and workload operations
there too now. So to piggy back on this need, this adds support for
listening on multiple namespaces if required.

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>